fix: DI-17 missed two more int32 index sites -- irc2.m mr2tr_ and irc.m mr2tr4_
Following up 9160c51. DI-17 (349a0e1) fixed int32 index saturation at two sites
in irc.m; a sweep for the same pattern found four in total.

irc.m:32699 mr2tr4_ -- FIXED. This is the most exposed of the four, not the
least: mr_T is the FULL recording (load_bin_T_) and viTime is the ABSOLUTE
viTime_spk, where the two sites DI-17 fixed index a loaded chunk. Past ~20 h the
indices wrap and the wrong samples are read, silently, since the results feed
medians. Two live callers, both reached with absolute times: the curation GUI
FigClust raw-waveform plot (:32662) and driftMatch_post_, i.e. post_merge_mode 3
(:4000).

irc2.m:1628 mr2tr_ -- FIXED. DI-17 was applied to irc.m only, so irc2 still had
the defect DI-17 was written to close. Reached from convert_mda.m,
convert_mda_ui.m, irc2phy.m, irc2klusters_v2.m and direct irc2 invocation;
irc.m itself calls irc2 only once (:34258, an irc2('describe') for benchmark
stats) and that path does not reach mr2tr_. double rather than int64 here for
the same reason as 9160c51 -- viTime is moved onto the GPU three lines above,
and gpuArray does not support int64 arithmetic.

irc.m:11757 recenter_spk_ -- NOT fixed, deliberately. Zero callers anywhere in
matlab/*.m, so it cannot execute. Its viTime_off at :11762 is also int32 and
would need the same treatment if the function is ever revived. Flagged rather
than changed blind, since nothing can exercise it.

Verified: both patched functions produce output IDENTICAL to the int32 they
replace at realistic lengths (mr2tr4_ full tensor, irc2 mr2tr_ both the plain
and the viSite branch); at 20 h the int32 arithmetic demonstrably differs while
double stays correct and in range; irc2 mr2tr_ runs on gpuArray and matches CPU,
and int64 still fails there. Both files parse with 0 errors.

+++ matlab/irc.m
@@ -32696,7 +32696,14 @@ function tr = mr2tr4_(mr_T, spkLim, viTime)
 nSpk = numel(viTime);
 
 viTime0 = [spkLim(1):spkLim(end)]';
-miRange = bsxfun(@plus, int32(viTime0), int32(viTime(:)'));
+% DI-17 fixed int32 index saturation in mr2tr_ and mn2tn_gpu_ but MISSED this site, which is
+% more exposed than either: mr_T here is the FULL recording (load_bin_T_ at :32612, :32711) and
+% viTime is the ABSOLUTE viTime_spk (:32619, :32736), where the two DI-17 sites index a loaded
+% chunk. int32 caps at 2.147e9 and 20h @ 30kHz is 2.16e9, so past ~20h the indices wrap and the
+% wrong samples are read -- silently, since these feed medians. double is exact to 2^53. See the
+% note in mr2tr_ for why double rather than int64. miRange is tiny here (<=2^12 spikes), so the
+% wider type costs nothing.
+miRange = bsxfun(@plus, double(viTime0), double(viTime(:)'));
 miRange = min(max(miRange, 1), nSamples);
 tr = mr_T(:, miRange(:));
 tr = reshape(tr, [nChans, numel(viTime0), numel(viTime)]);
